+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Todas las variables "white" ahora se llaman "bottom".
 - Todas las variables "rojo" ahora se llaman "top_color".
 - Todas las funciones que contenían "move_states" en su nombre, ahora contienen "pkmove_states".
+- Cambio de nombre de la función "get_peak_color" a "check_peak_ownership".
+- Cambio de nombre de la función "is_same_color" a "is_same_owner".
 
 ## [20190913a] - 2019-09-13 11:15
 ### Added